+++ catalyst/core/runner.py
@@ -319,7 +319,7 @@ class IRunner(ICallback, ILogger, ABC):
         """
         return None
 
-    def get_optimizer(self, model: Model, stage: str) -> Optimizer:
+    def get_optimizer(self, stage: str, model: Model) -> Optimizer:
         """Returns the optimizer for a given stage and model.
 
         Example::
@@ -337,7 +337,7 @@ class IRunner(ICallback, ILogger, ABC):
         """
         return None
 
-    def get_scheduler(self, optimizer: Optimizer, stage: str) -> Scheduler:
+    def get_scheduler(self, stage: str, optimizer: Optimizer) -> Scheduler:
         """Returns the scheduler for a given stage and optimizer.
 
         Example::
