+++ setup.py
@@ -88,20 +88,20 @@ class UploadCommand(Command):
 extras = {
     "cv": load_requirements("requirements/requirements-cv.txt"),
     "dev": load_requirements("requirements/requirements-dev.txt"),
-    "ecosystem": load_requirements("requirements/requirements-ecosystem.txt"),
-    "log": load_requirements("requirements/requirements-log.txt"),
+    # "ecosystem": load_requirements("requirements/requirements-ecosystem.txt"),
+    # "log": load_requirements("requirements/requirements-log.txt"),
     "ml": load_requirements("requirements/requirements-ml.txt"),
-    "nlp": load_requirements("requirements/requirements-nlp.txt"),
-    "tune": load_requirements("requirements/requirements-optuna.txt"),
+    # "nlp": load_requirements("requirements/requirements-nlp.txt"),
+    "optuna": load_requirements("requirements/requirements-optuna.txt"),
 }
-extras["contrib"] = (
-    extras["cv"]
-    + extras["ecosystem"]
-    + extras["log"]
-    + extras["ml"]
-    + extras["nlp"]
-    + extras["tune"]
-)
+# extras["contrib"] = (
+#     extras["cv"]
+#     + extras["ecosystem"]
+#     + extras["log"]
+#     + extras["ml"]
+#     + extras["nlp"]
+#     + extras["tune"]
+# )
 
 # Meta dependency groups.
 all_deps = []
